+++ libs/mcp/server-tools/src/lib/register-all-mcp-tools.ts
@@ -6,9 +6,9 @@ import { RegisterMCPTool_OpenDatasetsInENVI } from './tools/envi/register-mcp-to
 import { RegisterMCPTool_QueryDatasetWithENVI } from './tools/envi/register-mcp-tool-query-dataset-with-envi';
 import { RegisterMCPTool_SaveENVIToolWorkflow } from './tools/envi/register-mcp-tool-save-envi-tool-workflow';
 import { RegisterMCPTool_TakeENVIScreenshot } from './tools/envi/register-mcp-tool-take-envi-screenshot';
+import { RegisterMCPTool_ControlIDLDebugger } from './tools/idl/register-mcp-tool-control-idl-debugger';
 import { RegisterMCPTool_CreateIDLNotebook } from './tools/idl/register-mcp-tool-create-idl-notebook';
-import { RegisterMCPTool_InspectIDLState } from './tools/idl/register-mcp-tool-inspect-idl-state';
-import { RegisterMCPTool_ManageIDLDebugger } from './tools/idl/register-mcp-tool-manage-idl-debugger';
+import { RegisterMCPTool_GetIDLState } from './tools/idl/register-mcp-tool-get-idl-state';
 import { RegisterMCPTool_QueryIDLSession } from './tools/idl/register-mcp-tool-query-idl-session';
 import { RegisterMCPTool_RunIDLCode } from './tools/idl/register-mcp-tool-run-idl-code';
 import { RegisterMCPTool_RunIDLFile } from './tools/idl/register-mcp-tool-run-idl-file';
@@ -61,8 +61,8 @@ export function RegisterAllMCPTools(isEnviInstalled: boolean) {
   RegisterMCPTool_CreateIDLNotebook(server);
   RegisterMCPTool_RunIDLCode(server);
   RegisterMCPTool_RunIDLFile(server);
-  RegisterMCPTool_InspectIDLState(server);
-  RegisterMCPTool_ManageIDLDebugger(server);
+  RegisterMCPTool_GetIDLState(server);
+  RegisterMCPTool_ControlIDLDebugger(server);
   RegisterMCPTool_QueryIDLSession(server);
 
   /**