+++ libs/types/mcp/src/lib/idl/mcp-tool-get-idl-state.interface.ts
@@ -1,10 +1,10 @@
 import { IMCPToolIDL_BaseResponse } from '../mcp-base-response.interface';
 
 // name for server.
-export type MCPTool_InspectIDLState = 'inspect-idl-state';
+export type MCPTool_GetIDLState = 'get-idl-state';
 
 // list of actions.
-export type InspectIDLStateAction =
+export type GetIDLStateAction =
   | 'get-coverage'
   | 'get-errors'
   | 'get-info'
@@ -12,14 +12,14 @@ export type InspectIDLStateAction =
   | 'get-stack'
   | 'get-variables';
 
-// Parameters for inspecting IDL state
-export interface MCPToolParams_InspectIDLState {
-  action: InspectIDLStateAction;
+// Parameters for getting IDL state
+export interface MCPToolParams_GetIDLState {
+  action: GetIDLStateAction;
   // File path required for get-coverage
   file?: string;
   // defaults to current frame (0)
   frameId?: number;
 }
 
-// response type for inspecting IDL state
-export type MCPToolResponse_InspectIDLState = IMCPToolIDL_BaseResponse<any>;
+// response type for getting IDL state
+export type MCPToolResponse_GetIDLState = IMCPToolIDL_BaseResponse<any>;